break out rdnsquerier_testutils.go
